NO_ISSUE: Unify app providers configuration
