+++ projects/terminal-core-lib/src/terminal-providers/terminal-application-service-worker.providers.ts
@@ -1,0 +1,22 @@
+import {EnvironmentProviders} from '@angular/core';
+import {
+  provideServiceWorker,
+  SwRegistrationOptions
+} from '@angular/service-worker';
+
+export interface TerminalApplicationServiceWorkerOptions {
+  enabled: boolean;
+  registrationStrategy?: SwRegistrationOptions['registrationStrategy'];
+  script?: string;
+}
+
+export function provideTerminalServiceWorker(
+  options: TerminalApplicationServiceWorkerOptions
+): EnvironmentProviders[] {
+  return [
+    provideServiceWorker(options.script ?? 'ngsw-worker.js', {
+      enabled: options.enabled,
+      registrationStrategy: options.registrationStrategy ?? 'registerWithDelay:60000',
+    })
+  ];
+}
